+++ dbms/src/Parsers/ASTQueryWithOnCluster.cpp
@@ -4,6 +4,7 @@
 #include <Parsers/ExpressionElementParsers.h>
 #include <Parsers/ASTIdentifier.h>
 #include <Parsers/ASTLiteral.h>
+#include <Parsers/parseIdentifierOrStringLiteral.h>
 #include <Common/typeid_cast.h>
 #include <Interpreters/evaluateConstantExpression.h>
