+++ dbms/src/Parsers/parseIdentifierOrStringLiteral.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <Parsers/IParser.h>
+
+namespace DB
+{
+
+/** Parses a name of an object which could be written in 3 forms:
+  * name, `name` or 'name' */
+bool parseIdentifierOrStringLiteral(IParser::Pos & pos, Expected & expected, String & result);
+
+}
